+++ familyhelper-project-sdk/src/main/java/com/dwarfeng/familyhelper/project/sdk/bean/entity/WebInputTask.java
@@ -5,10 +5,12 @@ import com.dwarfeng.familyhelper.project.sdk.util.Constraints;
 import com.dwarfeng.familyhelper.project.stack.bean.entity.Task;
 import com.dwarfeng.subgrade.sdk.bean.key.WebInputLongIdKey;
 import com.dwarfeng.subgrade.stack.bean.Bean;
+import org.checkerframework.checker.index.qual.Positive;
 import org.hibernate.validator.constraints.Length;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Null;
+import javax.validation.constraints.PositiveOrZero;
 import java.util.Date;
 import java.util.Objects;
 
@@ -75,9 +77,11 @@ public class WebInputTask implements Bean {
     private Date finishedDate;
 
     @JSONField(name = "total_mission_count")
+    @Positive
     private int totalMissionCount;
 
     @JSONField(name = "finished_mission_count")
+    @PositiveOrZero
     private int finishedMissionCount;
 
     public WebInputTask() {